ENH: Re-organize lessons
Re-organize lessons: re-group the local orientation reconstruction and
tractography lessons as different subsections of their topics, and provide an
overview/introduction section for them.

Update the lessons table accordingly in the `README`.

Fix the links to the lessons pages in the `README`.

Take advantage of the commit to update the `About the Lesson` section in the
`README` to match the actual contents.